+++ pro/wiretap/src/examples/kotlin/wiretap/WiretapRunner.kt
@@ -8,10 +8,10 @@ import org.http4k.server.asServer
 import org.http4k.server.uri
 
 fun main() {
-//    val wiretap = HttpApp()
+    val wiretap = HttpApp()
 //    val wiretap = HttpAppWithOtelTracing()
 //    val wiretap = McpApp()
-    val wiretap = McpServer()
+//    val wiretap = McpServer()
 //    val wiretap = McpServerWithOtel()
 //    val wiretap = OpenApiApp()
 //    val wiretap = ExternalMcpServer()
@@ -22,9 +22,9 @@ fun main() {
 
     println("started ${server.uri().path("_wiretap")}")
 
-//    (0..5).forEach {
-//        JavaHttpClient()(Request(GET, server.uri().path(it.toString())))
-//    }
+    (0..5).forEach {
+        JavaHttpClient()(Request(GET, server.uri().path(it.toString())))
+    }
 
 }