+++ pro/wiretap/src/examples/kotlin/wiretap/examples/HttpApp.kt
@@ -6,7 +6,7 @@ import org.http4k.core.Response
 import org.http4k.core.Status.Companion.OK
 import org.http4k.lens.contentType
 
-fun HttpApp() = { request: Request ->
-    Response(OK).headers(request.headers)
+fun HttpApp() = { req: Request ->
+    Response(OK).headers(req.headers)
         .contentType(ContentType.APPLICATION_JSON).body("""{"hello":"world"}""")
 }